+++ src/Enums/PaymentStatus.php
@@ -11,6 +11,7 @@ namespace Subtain\LaravelPayments\Enums;
 enum PaymentStatus: string
 {
     case PENDING = 'pending';
+    case PROCESSING = 'processing';
     case PAID = 'paid';
     case FAILED = 'failed';
     case CANCELLED = 'cancelled';